+++ Open-S-Collab/projects/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class ProjectsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'projects'
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "projects"
