+++ Open-S-Collab/projects/models.py
@@ -7,4 +7,12 @@ class Project(models.Model):
     body = models.TextField(max_length=5000)
     image = models.ImageField(null=True, blank=True)
     created = models.DateTimeField(auto_now_add=True)
+    modified = models.DateTimeField(auto_now=True)
     owner = models.ForeignKey(Profile, on_delete=models.CASCADE)
+
+    class Meta:
+        verbose_name = "project"
+        verbose_name_plural = "projects"
+
+    def __str__(self):
+        return str(self.title)
