+++ Open-S-Collab/users/serializers.py
@@ -13,9 +13,10 @@ class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ["first_name", "email", "username", "password"]
-        extra_kwargs = {'password':{'write_only': True}}
+        extra_kwargs = {"password": {"write_only": True}}
+
     def create(self, validated_data):
-        password = validated_data.pop('password', None)
+        password = validated_data.pop("password", None)
         instance = self.Meta.model(**validated_data)
         if password is not None:
             instance.set_password(password)
